requirements.txt
Flask==3.1.1
Flask-Session==0.8.0
bcrypt==4.3.0
dotenv==0.9.9
python-dotenv==1.1.1
mysql-connector-python